Pull previewing a design into its own component
